# Code speedup due to faster einsum replacement