@oseiskar 2023-09-09
Worked fine. Merged and released in [1.0.3](https://github.com/oseiskar/simdkalman/releases/tag/1.0.3). Thanks again!